+++ include/osl1/convex_relation_ptr.h
@@ -32,6 +32,11 @@
 typedef osl1_convex_relation_t* osl1_convex_relation_ptr_t;
 
 
+#if defined(__cplusplus)
+extern "C"
+{
+#endif
+
 // Create & destroy
 osl1_convex_relation_ptr_t osl1_convex_relation_ptr_create();
 void osl1_convex_relation_ptr_destroy(
@@ -65,5 +70,9 @@ bool osl1_convex_relation_ptr_equal(const osl1_convex_relation_ptr_t* const a,
 // Conversion
 gho_string_t osl1_convex_relation_ptr_to_string(
                   const osl1_convex_relation_ptr_t* const convex_relation_ptr);
+
+#if defined(__cplusplus)
+}
+#endif
 
 #endif
