+++ include/osl1/implementation/print/convex_relation.h
@@ -17,11 +17,57 @@
 
 #include <gho/output.h>
 #include <gho/c_str.h>
+#include <gho/vector/uint.h>
 
-#include "../include/osl1/print.h"
+#include "../../print.h"
+#include "../../extension.h"
 
 
 // Convex relation index name
+
+static inline
+gho_string_t osl1_convex_relation_name_column(
+                            const osl1_convex_relation_t* const convex_relation,
+                            const size_t j,
+                            const osl1_scop_t* const scop,
+                            const osl1_statement_t* const statement);
+static inline
+gho_string_t osl1_convex_relation_name_output_dim(
+                            const osl1_convex_relation_t* const convex_relation,
+                            const size_t i_output_dim,
+                            const osl1_scop_t* const scop,
+                            const osl1_statement_t* const statement);
+static inline
+gho_string_t osl1_convex_relation_name_input_dim(
+                            const osl1_convex_relation_t* const convex_relation,
+                            const size_t i_input_dim,
+                            const osl1_scop_t* const scop,
+                            const osl1_statement_t* const statement);
+static inline
+gho_string_t osl1_convex_relation_name_local_dim(
+                            const osl1_convex_relation_t* const convex_relation,
+                            const size_t i_local_dim,
+                            const osl1_scop_t* const scop,
+                            const osl1_statement_t* const statement);
+static inline
+gho_string_t osl1_convex_relation_name_parameter(
+                            const osl1_convex_relation_t* const convex_relation,
+                            const size_t i_parameter,
+                            const osl1_scop_t* const scop,
+                            const osl1_statement_t* const statement);
+static inline
+void osl1_convex_relation_fprint_name(FILE * file,
+                                      const gho_string_t* const name,
+                                      const gho_uint column_size);
+static inline
+void osl1_convex_relation_sprint_name(char** c_str,
+                                      const gho_string_t* const name,
+                                      const gho_uint column_size);
+static inline
+gho_vector_uint_t osl1_convex_relation_column_sizes(
+                            const osl1_convex_relation_t* const convex_relation,
+                            const osl1_scop_t* const scop,
+                            const osl1_statement_t* const statement);
 
 /**
  * \brief Return name of a dimension
@@ -754,391 +800,3 @@ void osl1_convex_relation_sprint(char** c_str,
                           const osl1_convex_relation_t* const convex_relation) {
   osl1_convex_relation_sprinti(c_str, convex_relation, 0);
 }
-
-// Relation
-
-/**
- * \brief Print an osl1_relation_t in a file with indentation
- * \param[in] file      A C file
- * \param[in] relation  An osl1_relation_t
- * \param[in] indent    Indentation (number of spaces)
- * \param[in] scop      An osl1_scop_t (can be NULL)
- * \param[in] statement An osl1_statement_t (can be NULL)
- */
-void osl1_relation_fprinti_openscop(FILE* file,
-                                    const osl1_relation_t* const relation,
-                                    const unsigned int indent,
-                                    const osl1_scop_t* const scop,
-                                    const osl1_statement_t* const statement) {
-  
-  gho_c_str_fprinti(file, "# Number of convex relations\n", indent);
-  gho_size_t_fprinti(file, &relation->size, indent);
-  gho_c_str_fprint(file, "\n\n");
-  
-  for (size_t i = 0; i < relation->size; ++i) {
-    gho_c_str_fprinti(file, "# Convex relation ", indent);
-    const size_t tmp = i + 1;
-    gho_size_t_fprint(file, &tmp);
-    gho_c_str_fprint(file, " / ");
-    gho_size_t_fprint(file, &relation->size);
-    gho_c_str_fprint(file, "\n\n");
-    osl1_convex_relation_fprinti_openscop(file, &relation->array[i], indent,
-                                          scop, statement);
-    if (i != relation->size - 1) { gho_c_str_fprint(file, "\n"); }
-  }
-}
-
-/**
- * \brief Print an osl1_relation_t in a C string with indentation
- * \param[in] c_str     A C string
- * \param[in] relation  An osl1_relation_t
- * \param[in] indent    Indentation (number of spaces)
- * \param[in] scop      An osl1_scop_t (can be NULL)
- * \param[in] statement An osl1_statement_t (can be NULL)
- */
-void osl1_relation_sprinti_openscop(char** c_str,
-                                    const osl1_relation_t* const relation,
-                                    const unsigned int indent,
-                                    const osl1_scop_t* const scop,
-                                    const osl1_statement_t* const statement) {
-  
-  gho_c_str_sprinti(c_str, "# Number of convex relations\n", indent);
-  gho_size_t_sprinti(c_str, &relation->size, indent);
-  gho_c_str_sprint(c_str, "\n\n");
-  
-  for (size_t i = 0; i < relation->size; ++i) {
-    gho_c_str_sprinti(c_str, "# Convex relation ", indent);
-    const size_t tmp = i + 1;
-    gho_size_t_sprint(c_str, &tmp);
-    gho_c_str_sprint(c_str, " / ");
-    gho_size_t_sprint(c_str, &relation->size);
-    gho_c_str_sprint(c_str, "\n\n");
-    osl1_convex_relation_sprinti_openscop(c_str, &relation->array[i], indent,
-                                          scop, statement);
-    if (i != relation->size - 1) { gho_c_str_sprint(c_str, "\n"); }
-  }
-}
-
-// Vector of relations
-
-/**
- * \brief Print an osl1_vector_relation_t in a file with indentation
- * \param[in] file      A C file
- * \param[in] relations An osl1_vector_relation_t
- * \param[in] indent    Indentation (number of spaces)
- * \param[in] scop      An osl1_scop_t (can be NULL)
- * \param[in] statement An osl1_statement_t (can be NULL)
- */
-void osl1_vector_relation_fprinti_openscop(FILE* file,
-                                  const osl1_vector_relation_t* const relations,
-                                  const unsigned int indent,
-                                  const osl1_scop_t* const scop,
-                                  const osl1_statement_t* const statement) {
-  
-  for (size_t i = 0; i < relations->size; ++i) {
-    gho_c_str_fprinti(file, "# Relation ", indent);
-    const size_t tmp = i + 1;
-    gho_size_t_fprint(file, &tmp);
-    gho_c_str_fprint(file, " / ");
-    gho_size_t_fprint(file, &relations->size);
-    gho_c_str_fprint(file, "\n\n");
-    osl1_relation_fprinti_openscop(file, &relations->array[i], indent,
-                                   scop, statement);
-    if (i != relations->size - 1) { gho_c_str_fprint(file, "\n"); }
-  }
-}
-
-/**
- * \brief Print an osl1_vector_relation_t in a C string with indentation
- * \param[in] c_str     A C string
- * \param[in] relations An osl1_vector_relation_t
- * \param[in] indent    Indentation (number of spaces)
- * \param[in] scop      An osl1_scop_t (can be NULL)
- * \param[in] statement An osl1_statement_t (can be NULL)
- */
-void osl1_vector_relation_sprinti_openscop(char** c_str,
-                                  const osl1_vector_relation_t* const relations,
-                                  const unsigned int indent,
-                                  const osl1_scop_t* const scop,
-                                  const osl1_statement_t* const statement) {
-  
-  for (size_t i = 0; i < relations->size; ++i) {
-    gho_c_str_sprinti(c_str, "# Relation ", indent);
-    const size_t tmp = i + 1;
-    gho_size_t_sprint(c_str, &tmp);
-    gho_c_str_sprint(c_str, " / ");
-    gho_size_t_sprint(c_str, &relations->size);
-    gho_c_str_sprint(c_str, "\n\n");
-    osl1_relation_sprinti_openscop(c_str, &relations->array[i], indent,
-                                   scop, statement);
-    if (i != relations->size - 1) { gho_c_str_sprint(c_str, "\n"); }
-  }
-}
-
-// Statement
-
-/**
- * \brief Print an osl1_statement_t in a file with indentation
- * \param[in] file      A C file
- * \param[in] statement An osl1_statement_t
- * \param[in] indent    Indentation (number of spaces)
- * \param[in] scop      An osl1_scop_t (can be NULL)
- */
-void osl1_statement_fprinti_openscop(FILE* file,
-                                     const osl1_statement_t* const statement,
-                                     const unsigned int indent,
-                                     const osl1_scop_t* const scop) {
-  
-  gho_c_str_fprinti(file, "# Id = ", indent);
-  gho_size_t_fprint(file, &statement->id);
-  gho_c_str_fprint(file, "\n\n");
-  
-  gho_c_str_fprinti(file, "# Domain\n", indent);
-  gho_c_str_fprinti(file, "# ------\n\n", indent);
-  osl1_relation_fprinti_openscop(file, &statement->domain, indent,
-                                 scop, statement);
-  gho_c_str_fprint(file, "\n\n");
-  
-  gho_c_str_fprinti(file, "# Scattering\n", indent);
-  gho_c_str_fprinti(file, "# ----------\n\n", indent);
-  osl1_relation_fprinti_openscop(file, &statement->scattering, indent,
-                                 scop, statement);
-  gho_c_str_fprint(file, "\n\n");
-  
-  gho_c_str_fprinti(file, "# Number of accesses relations (read)\n", indent);
-  gho_size_t_fprinti(file, &statement->accesses_read.size, indent);
-  gho_c_str_fprint(file, "\n\n");
-  if (statement->accesses_read.size != 0) {
-    gho_c_str_fprinti(file, "# Accesses (read)\n", indent);
-    gho_c_str_fprinti(file, "# ---------------\n\n", indent);
-    osl1_vector_relation_fprinti_openscop(file, &statement->accesses_read,
-                                          indent, scop, statement);
-    gho_c_str_fprint(file, "\n\n");
-  }
-  
-  gho_c_str_fprinti(file, "# Number of accesses relations (write)\n", indent);
-  gho_size_t_fprinti(file, &statement->accesses_write.size, indent);
-  gho_c_str_fprint(file, "\n\n");
-  if (statement->accesses_write.size != 0) {
-    gho_c_str_fprinti(file, "# Accesses (write)\n", indent);
-    gho_c_str_fprinti(file, "# ----------------\n\n", indent);
-    osl1_vector_relation_fprinti_openscop(file, &statement->accesses_write,
-                                          indent, scop, statement);
-    gho_c_str_fprint(file, "\n\n");
-  }
-  
-  gho_c_str_fprinti(file, "# Number of statement extensions\n", indent);
-  gho_size_t_fprinti(file, &statement->extensions.size, indent);
-  gho_c_str_fprint(file, "\n\n");
-  gho_c_str_fprinti(file, "# Statement extensions\n", indent);
-  gho_c_str_fprinti(file, "# --------------------\n\n", indent);
-  osl1_extensions_fprinti_openscop(file, &statement->extensions, indent);
-}
-
-/**
- * \brief Print an osl1_statement_t in a C string with indentation
- * \param[in] c_str     A C string
- * \param[in] statement An osl1_statement_t
- * \param[in] indent    Indentation (number of spaces)
- * \param[in] scop      An osl1_scop_t (can be NULL)
- */
-void osl1_statement_sprinti_openscop(char** c_str,
-                                     const osl1_statement_t* const statement,
-                                     const unsigned int indent,
-                                     const osl1_scop_t* const scop) {
-  
-  gho_c_str_sprinti(c_str, "# Id = ", indent);
-  gho_size_t_sprint(c_str, &statement->id);
-  gho_c_str_sprint(c_str, "\n\n");
-  
-  gho_c_str_sprinti(c_str, "# Domain\n", indent);
-  gho_c_str_sprinti(c_str, "# ------\n\n", indent);
-  osl1_relation_sprinti_openscop(c_str, &statement->domain, indent,
-                                 scop, statement);
-  gho_c_str_sprint(c_str, "\n\n");
-  
-  gho_c_str_sprinti(c_str, "# Scattering\n", indent);
-  gho_c_str_sprinti(c_str, "# ----------\n\n", indent);
-  osl1_relation_sprinti_openscop(c_str, &statement->scattering, indent,
-                                 scop, statement);
-  gho_c_str_sprint(c_str, "\n\n");
-  
-  gho_c_str_sprinti(c_str, "# Number of accesses relations (read)\n", indent);
-  gho_size_t_sprinti(c_str, &statement->accesses_read.size, indent);
-  gho_c_str_sprint(c_str, "\n\n");
-  if (statement->accesses_read.size != 0) {
-    gho_c_str_sprinti(c_str, "# Accesses (read)\n", indent);
-    gho_c_str_sprinti(c_str, "# ---------------\n\n", indent);
-    osl1_vector_relation_sprinti_openscop(c_str, &statement->accesses_read,
-                                          indent, scop, statement);
-    gho_c_str_sprint(c_str, "\n\n");
-  }
-  
-  gho_c_str_sprinti(c_str, "# Number of accesses relations (write)\n", indent);
-  gho_size_t_sprinti(c_str, &statement->accesses_write.size, indent);
-  gho_c_str_sprint(c_str, "\n\n");
-  if (statement->accesses_write.size != 0) {
-    gho_c_str_sprinti(c_str, "# Accesses (write)\n", indent);
-    gho_c_str_sprinti(c_str, "# ----------------\n\n", indent);
-    osl1_vector_relation_sprinti_openscop(c_str, &statement->accesses_write,
-                                          indent, scop, statement);
-    gho_c_str_sprint(c_str, "\n\n");
-  }
-  
-  gho_c_str_sprinti(c_str, "# Number of statement extensions\n", indent);
-  gho_size_t_sprinti(c_str, &statement->extensions.size, indent);
-  gho_c_str_sprint(c_str, "\n\n");
-  gho_c_str_sprinti(c_str, "# Statement extensions\n", indent);
-  gho_c_str_sprinti(c_str, "# --------------------\n\n", indent);
-  osl1_extensions_sprinti_openscop(c_str, &statement->extensions, indent);
-}
-
-// Vector of statements
-
-/**
- * \brief Print an osl1_vector_statement_t in a file with indentation
- * \param[in] file       A C file
- * \param[in] statements An osl1_vector_statement_t
- * \param[in] indent     Indentation (number of spaces)
- * \param[in] scop       An osl1_scop_t (can be NULL)
- */
-void osl1_vector_statement_fprinti_openscop(FILE* file,
-                                const osl1_vector_statement_t* const statements,
-                                const unsigned int indent,
-                                const osl1_scop_t* const scop) {
-  gho_c_str_fprinti(file, "# Number of statements\n", indent);
-  gho_size_t_fprinti(file, &statements->size, indent);
-  gho_c_str_fprint(file, "\n\n");
-  
-  for (size_t i = 0; i < statements->size; ++i) {
-    gho_c_str_fprint(file, "# Statement ");
-    const size_t tmp = i + 1;
-    gho_size_t_fprint(file, &tmp);
-    gho_c_str_fprint(file, " / ");
-    gho_size_t_fprint(file, &statements->size);
-    gho_c_str_fprint(file, "\n");
-    gho_c_str_fprint(file, "# ---------------\n\n");
-    osl1_statement_fprinti_openscop(file, &statements->array[i], indent + 4,
-                                    scop);
-    if (i != statements->size - 1) { gho_c_str_fprint(file, "\n\n"); }
-  }
-}
-
-/**
- * \brief Print an osl1_vector_statement_t in a C string with indentation
- * \param[in] c_str      A C string
- * \param[in] statements An osl1_vector_statement_t
- * \param[in] indent     Indentation (number of spaces)
- * \param[in] scop       An osl1_scop_t (can be NULL)
- */
-void osl1_vector_statement_sprinti_openscop(char** c_str,
-                                const osl1_vector_statement_t* const statements,
-                                const unsigned int indent,
-                                const osl1_scop_t* const scop) {
-  gho_c_str_sprinti(c_str, "# Number of statements\n", indent);
-  gho_size_t_sprinti(c_str, &statements->size, indent);
-  gho_c_str_sprint(c_str, "\n\n");
-  
-  for (size_t i = 0; i < statements->size; ++i) {
-    gho_c_str_sprint(c_str, "# Statement ");
-    const size_t tmp = i + 1;
-    gho_size_t_sprint(c_str, &tmp);
-    gho_c_str_sprint(c_str, " / ");
-    gho_size_t_sprint(c_str, &statements->size);
-    gho_c_str_sprint(c_str, "\n");
-    gho_c_str_sprint(c_str, "# ---------------\n\n");
-    osl1_statement_sprinti_openscop(c_str, &statements->array[i], indent + 4,
-                                    scop);
-    if (i != statements->size - 1) { gho_c_str_sprint(c_str, "\n\n"); }
-  }
-}
-
-// SCoP
-
-/**
- * \brief Print an osl1_scop_t in a file with indentation
- * \param[in] file   A C file
- * \param[in] scop   An osl1_scop_t
- * \param[in] indent Indentation (number of spaces)
- */
-void osl1_scop_fprinti_openscop(FILE* file,
-                                const osl1_scop_t* const scop,
-                                const unsigned int indent) {
-  
-  gho_c_str_fprinti(file, "<OpenScop1>\n", indent);
-  gho_c_str_fprint(file, "\n");
-  
-  gho_c_str_fprinti(file, "# Context\n", indent);
-  gho_c_str_fprinti(file, "# -------\n\n", indent);
-  osl1_relation_fprinti_openscop(file, &scop->context, indent, scop, NULL);
-  gho_c_str_fprint(file, "\n\n");
-  
-  gho_c_str_fprinti(file, "# Parameters\n", indent);
-  gho_c_str_fprinti(file, "# ----------\n", indent);
-  gho_size_t_fprinti(file, &scop->parameters.size, indent);
-  gho_c_str_fprint(file, "\n");
-  for (size_t i = 0; i < scop->parameters.size; ++i) {
-    gho_string_fprinti(file, &scop->parameters.array[i], indent);
-    gho_c_str_fprint(file, " ");
-  }
-  gho_c_str_fprint(file, "\n\n");
-  
-  osl1_vector_statement_fprinti_openscop(file, &scop->statements, indent,
-                                         scop);
-  gho_c_str_fprint(file, "\n\n");
-  
-  gho_c_str_fprinti(file, "# Number of SCoP extensions\n", indent);
-  gho_size_t_fprinti(file, &scop->extensions.size, indent);
-  gho_c_str_fprint(file, "\n\n");
-  gho_c_str_fprinti(file, "# SCoP extensions\n", indent);
-  gho_c_str_fprinti(file, "# ---------------\n\n", indent);
-  osl1_extensions_fprinti_openscop(file, &scop->extensions, indent);
-  gho_c_str_fprint(file, "\n\n");
-  
-  gho_c_str_fprinti(file, "</OpenScop1>", indent);
-}
-
-/**
- * \brief Print an osl1_scop_t in a C string with indentation
- * \param[in] c_str  A C string
- * \param[in] scop   An osl1_scop_t
- * \param[in] indent Indentation (number of spaces)
- */
-void osl1_scop_sprinti_openscop(char** c_str,
-                                const osl1_scop_t* const scop,
-                                const unsigned int indent) {
-  
-  gho_c_str_sprinti(c_str, "<OpenScop1>\n", indent);
-  gho_c_str_sprint(c_str, "\n");
-  
-  gho_c_str_sprinti(c_str, "# Context\n", indent);
-  gho_c_str_sprinti(c_str, "# -------\n\n", indent);
-  osl1_relation_sprinti_openscop(c_str, &scop->context, indent, scop, NULL);
-  gho_c_str_sprint(c_str, "\n\n");
-  
-  gho_c_str_sprinti(c_str, "# Parameters\n", indent);
-  gho_c_str_sprinti(c_str, "# ----------\n", indent);
-  gho_size_t_sprinti(c_str, &scop->parameters.size, indent);
-  gho_c_str_sprint(c_str, "\n");
-  for (size_t i = 0; i < scop->parameters.size; ++i) {
-    gho_string_sprinti(c_str, &scop->parameters.array[i], indent);
-    gho_c_str_sprint(c_str, " ");
-  }
-  gho_c_str_sprint(c_str, "\n\n");
-  
-  osl1_vector_statement_sprinti_openscop(c_str, &scop->statements, indent,
-                                         scop);
-  gho_c_str_sprint(c_str, "\n\n");
-  
-  gho_c_str_sprinti(c_str, "# Number of SCoP extensions\n", indent);
-  gho_size_t_sprinti(c_str, &scop->extensions.size, indent);
-  gho_c_str_sprint(c_str, "\n\n");
-  gho_c_str_sprinti(c_str, "# SCoP extensions\n", indent);
-  gho_c_str_sprinti(c_str, "# ---------------\n\n", indent);
-  osl1_extensions_sprinti_openscop(c_str, &scop->extensions, indent);
-  gho_c_str_sprint(c_str, "\n\n");
-  
-  gho_c_str_sprinti(c_str, "</OpenScop1>", indent);
-}
-
-// Vector of SCoP
