+++ include/osl1/implementation/vector/statement.h
@@ -13,12 +13,6 @@
 // limitations under the License.
 
 
-#include <gho/int/size_t.h>
-
-#include "../include/osl1/vector/statement.h"
-#include "../include/osl1/input.h"
-
-
 // Input
 
 /**
