+++ include/osl1/relation.h
@@ -51,10 +51,15 @@
 
 
 // Input
+static inline
 osl1_relation_t osl1_relation_fread(FILE* file);
+static inline
 osl1_relation_t osl1_relation_sread(const char** c_str);
 
 // Output dimensions
 size_t osl1_relation_max_nb_output_dim(const osl1_relation_t* const relation);
+
+
+#include "implementation/relation.h"
 
 #endif
