+++ include/osl1/statement.h
@@ -25,13 +25,6 @@
 
 #include "relation.h"
 #include "vector/relation.h"
-#include "extension.h"
-
-
-#if defined(__cplusplus)
-extern "C"
-{
-#endif
 
 
 /**
@@ -60,43 +53,55 @@ typedef struct {
   
 } osl1_statement_t;
 
-// Id
-size_t osl1_statement_new_id();
 
 // Create & destroy
+static inline
 osl1_statement_t osl1_statement_create();
+static inline
 void osl1_statement_destroy(osl1_statement_t* statement);
+static inline
 void osl1_statement_erase(osl1_statement_t* statement);
 
 // Output
+static inline
 void osl1_statement_fprinti(FILE* file,
                             const osl1_statement_t* const statement,
                             const unsigned int indent);
+static inline
 void osl1_statement_fprint(FILE* file,
                            const osl1_statement_t* const statement);
+static inline
 void osl1_statement_print(const osl1_statement_t* const statement);
+static inline
 void osl1_statement_sprinti(char** c_str,
                             const osl1_statement_t* const statement,
                             const unsigned int indent);
+static inline
 void osl1_statement_sprint(char** c_str,
                            const osl1_statement_t* const statement);
 // Input
+static inline
 osl1_statement_t osl1_statement_fread(FILE* file);
+static inline
 osl1_statement_t osl1_statement_sread(const char** c_str);
 
 // Copy & comparisons
+static inline
 osl1_statement_t osl1_statement_copy(const osl1_statement_t* const statement);
+static inline
 void osl1_statement_copy_(const osl1_statement_t* const statement,
                           osl1_statement_t* copy);
+static inline
 bool osl1_statement_equal(const osl1_statement_t* const a,
                           const osl1_statement_t* const b);
 
 // Conversion
+static inline
 gho_string_t osl1_statement_to_string(const osl1_statement_t* const statement);
+static inline
 gho_any_t osl1_statement_to_any(const osl1_statement_t* const statement);
 
-#if defined(__cplusplus)
-}
-#endif
+
+#include "implementation/statement.h"
 
 #endif
