+++ src/osl1/extension/body.c
@@ -274,24 +274,3 @@ gho_string_t osl1_extension_body_to_string(
   osl1_extension_body_sprint(&r.c_str, body);
   return r;
 }
-
-/**
- * \brief Convert an osl1_extension_body into a gho_any
- * \param[in] body An osl1_extension_body
- * \return the gho_any from the osl1_extension_body
- */
-gho_any_t osl1_extension_body_to_any(const osl1_extension_body_t* const body) {
-  gho_any_t r = gho_any_create();
-  r.type = GHO_TYPE_OSL_EXTENSION_BODY;
-  r.size_of_struct = sizeof(osl1_extension_body_t);
-  osl1_extension_body_t* p = gho_alloc(osl1_extension_body_t);
-  osl1_extension_body_copy_(body, p);
-  r.any = p; p = NULL;
-  r.destroy_fct = (gho_destroy_fct_t)osl1_extension_body_destroy;
-  r.fprinti_fct = (gho_fprinti_fct_t)osl1_extension_body_fprinti;
-  r.sprinti_fct = (gho_sprinti_fct_t)osl1_extension_body_sprinti;
-  r.copy_fct = (gho_copy_fct_t)osl1_extension_body_copy_;
-  r.equal_fct = (gho_equal_fct_t)osl1_extension_body_equal;
-  r.to_string_fct = (gho_to_string_fct_t)osl1_extension_body_to_string;
-  return r;
-}
