+++ src/osl1/extension/unknown.c
@@ -265,25 +265,3 @@ gho_string_t osl1_extension_unknown_to_string(
   osl1_extension_unknown_sprint(&s.c_str, unknown);
   return s;
 }
-
-/**
- * \brief Convert an osl1_extension_unknown into an osl1_any
- * \param[in] unknown An osl1_extension_unknown
- * \return the osl1_any from the osl1_extension_unknown
- */
-gho_any_t osl1_extension_unknown_to_any(
-                                const osl1_extension_unknown_t* const unknown) {
-  gho_any_t r = gho_any_create();
-  r.type = GHO_TYPE_OSL_EXTENSION_UNKNOWN;
-  r.size_of_struct = sizeof(osl1_extension_unknown_t);
-  osl1_extension_unknown_t* p = gho_alloc(osl1_extension_unknown_t);
-  osl1_extension_unknown_copy_(unknown, p);
-  r.any = p; p = NULL;
-  r.destroy_fct = (gho_destroy_fct_t)osl1_extension_unknown_destroy;
-  r.fprinti_fct = (gho_fprinti_fct_t)osl1_extension_unknown_fprinti;
-  r.sprinti_fct = (gho_sprinti_fct_t)osl1_extension_unknown_sprinti;
-  r.copy_fct = (gho_copy_fct_t)osl1_extension_unknown_copy_;
-  r.equal_fct = (gho_equal_fct_t)osl1_extension_unknown_equal;
-  r.to_string_fct = (gho_to_string_fct_t)osl1_extension_unknown_to_string;
-  return r;
-}
