--- tests/osl1/extension.c
@@ -1,38 +1,0 @@
-// Copyright © 2015 Inria, Written by Lénaïc Bagnères, lenaic.bagneres@inria.fr
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-
-#include <stdio.h>
-#include <stdlib.h>
-
-#include <osl1/extension.h>
-#include <gho/test.h>
-
-
-int main(int argc, char** argv) {
-  
-  printf("\n");
-  
-  if (argc > 1) { printf("%s: argv are ignored\n", argv[0]); }
-
-  int nb_fail = 0;
-  
-  
-  
-  printf("%s ", argv[0]);
-  printf("fails = %d\n", nb_fail);
-  printf("\n");
-  
-  return nb_fail;
-}
