+++ tests/osl1/vector_statement.c
@@ -17,6 +17,7 @@
 #include <stdlib.h>
 
 #include <osl1/vector/statement.h>
+#include <osl1/print.h>
 #include <gho/test.h>
